requirements-tests.txt
pytest
pytest-mock
pytest-cov
mock
pysmbus
